Updated DenseVectorChart to store the entire chart in a single parallel array. Adapted OpenCL parser to use DenseVectorChart storage and parse entirely in the GPU (without copying back and forth to host memory).
Improves OpenCL efficiency considerably (~4X?) but not dramatically.

Some additional documentation in OpenCL code.

+++ java/src/edu/ohsu/cslu/parser/SortAndScanCsrSpmvParser.java
@@ -129,7 +129,7 @@ public class SortAndScanCsrSpmvParser extends
 
             final int rightCellRightChildren = rightCell.maxRightChildIndex() - rightCell.offset() + 1;
 
-            final int[] nonTerminalIndices = chart.nonTerminalIndices;
+            final short[] nonTerminalIndices = chart.nonTerminalIndices;
             final float[] insideProbabilities = chart.insideProbabilities;
 
             for (int i = leftCell.minLeftChildIndex(); i <= leftCell.maxLeftChildIndex(); i++) {
